+++ taf/updater/updater_pipeline.py
@@ -322,11 +322,6 @@ class AuthenticationRepositoryUpdatePipeline(Pipeline):
                     self.should_update_auth_repos,
                 ),  # auth repo
                 (
-                    self.check_if_previous_update_partial,
-                    RunMode.ALL,
-                    self.should_run_step_default,
-                ),
-                (
                     self.validate_last_validated_commit,
                     RunMode.ALL,
                     self.should_update_auth_repos,
@@ -345,6 +340,11 @@ class AuthenticationRepositoryUpdatePipeline(Pipeline):
                     self.load_target_repositories,
                     RunMode.ALL,
                     self.should_run_if_not_synced,
+                ),
+                (
+                    self.check_if_previous_update_partial,
+                    RunMode.ALL,
+                    self.should_run_step_default,
                 ),
                 (
                     self.set_auth_commit_for_target_repos,
@@ -790,39 +790,36 @@ class AuthenticationRepositoryUpdatePipeline(Pipeline):
         """
 
         def _previous_update_partial() -> bool:
-            if not self.state.users_auth_repo.last_validated_commit:
-                return True
+            if self.operation == OperationType.CLONE:
+                return False
+
+            last_validated_data = self.state.users_auth_repo.last_validated_data
+            auth_repo_last_validated_commit = (
+                self.state.users_auth_repo.last_validated_commit
+            )
 
             # validate from last_validated_commit if last_validated_data does not exist
             # (if last_validated_data was deleted)
-            if not self.state.users_auth_repo.last_validated_data:
+            if not last_validated_data or not auth_repo_last_validated_commit:
                 return True
+
+            auth_repo_commit = last_validated_data.get(self.state.users_auth_repo.name)
             if (
-                self.state.users_auth_repo.last_validated_commit
-                != self.state.users_auth_repo.last_validated_data.get(
-                    self.state.users_auth_repo.name
-                )
+                auth_repo_commit is None
+                or auth_repo_commit != auth_repo_last_validated_commit
             ):
                 return True
 
-            last_validated_commits = set()
             for repo_name in self.state.users_target_repositories:
-                if self.state.users_auth_repo.last_validated_data:
-                    if repo_name not in self.state.users_auth_repo.last_validated_data:
-                        return True
-                    last_validated_commits.add(
-                        self.state.users_auth_repo.last_validated_data[repo_name]
-                    )
-            return len(last_validated_commits) > 1
+                if repo_name in self.excluded_target_names:
+                    continue
+                if repo_name not in last_validated_data:
+                    return True
+                if last_validated_data[repo_name] != auth_repo_commit:
+                    return True
+            return False
 
-        is_partial = _previous_update_partial()
-        self.state.is_partially_updated = is_partial
-        if is_partial:
-            self.state.last_validated_commit = (
-                self.state.users_auth_repo.last_validated_data.get(
-                    self.state.users_auth_repo.name
-                )
-            )
+        self.state.is_partially_updated = _previous_update_partial()
 
     def _check_if_target_repos_clean(self, target_repos, branches_per_repo):
         dirty_index_repos = []
@@ -2052,7 +2049,6 @@ but commit not on branch {current_branch}"
             last_validated_data[self.state.users_auth_repo.name] = last_commit.value
             self.state.users_auth_repo.set_last_validated_data(
                 last_validated_data,
-                set_last_validated_commit=not bool(self.excluded_target_names),
             )
 
             return self.state.update_status